+++ Unity/Assets/Scripts/Core/World/Module/EventSystem/InvokeAttribute.cs
@@ -2,9 +2,9 @@
 {
     public class InvokeAttribute: BaseAttribute
     {
-        public int Type { get; }
+        public long Type { get; }
 
-        public InvokeAttribute(int type = 0)
+        public InvokeAttribute(long type = 0)
         {
             this.Type = type;
         }
